+++ Backend/src/main/java/com/gestor_clinica_veterinaria/VeterinaryHospitalManager/Entity/Surgery.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Date;
 @Entity
@@ -24,7 +25,7 @@ public class Surgery {
     private String surgeryProcedure;
     private String observations;
     private String postSurgeryRecommendations;
-    private Double surgeryCost;
+    private BigDecimal surgeryCost;
 
 
   /*
